third_party/blink/web_tests/platform/win/external/wpt/clipboard-apis/clipboard-copy-selection-line-break.https_21-30-expected.txt

This is a testharness.js-based test.
[FAIL] test preformatted text with line breaks: "\\n\\r\\r"
  assert_equals: Check result from DataTransfer in paste event expected "First\\r\\n\\r\\rSecond" but got "First\\r\\n\\r\\n\\r\\nSecond"
[FAIL] test preformatted text with line breaks: "\\n\\n\\r"
  assert_equals: Check result from DataTransfer in paste event expected "First\\r\\n\\r\\n\\rSecond" but got "First\\r\\n\\r\\n\\r\\nSecond"
[FAIL] test preformatted text with line breaks: "\\r \\n"
  assert_equals: Check result from DataTransfer in paste event expected "First\\r \\r\\nSecond" but got "First\\r\\n \\r\\nSecond"
Harness: the test ran to completion.
